FMWK-559-read-after-parameter - added read after parameter to cloud storages
