+++ io/azure/blob/reader.go
@@ -41,57 +41,6 @@ type Reader struct {
 	prefix string
 }
 
-type options struct {
-	// path contains path to file or directory.
-	path string
-	// isDir flag describes what we have in path, file or directory.
-	isDir bool
-	// isRemovingFiles flag describes should we remove everything from backup folder or not.
-	isRemovingFiles bool
-	// validator contains files validator that is applied to files if isDir = true.
-	validator validator
-	// Concurrency defines the max number of concurrent uploads to be performed to upload the file.
-	// Each concurrent upload will create a buffer of size BlockSize.
-	uploadConcurrency int
-	// withNestedDir describes if we should check for if an object is a directory for read/write operations.
-	// When we stream files or delete files in folder, we skip directories. This flag will avoid skipping.
-	// Default: false
-	withNestedDir bool
-}
-
-type Opt func(*options)
-
-// WithDir adds directory to reading/writing files from/to.
-func WithDir(path string) Opt {
-	return func(r *options) {
-		r.path = path
-		r.isDir = true
-	}
-}
-
-// WithFile adds a file path to reading/writing from/to.
-func WithFile(path string) Opt {
-	return func(r *options) {
-		r.path = path
-		r.isDir = false
-	}
-}
-
-// WithValidator adds validator to Reader, so files will be validated before reading.
-// Is used only for Reader.
-func WithValidator(v validator) Opt {
-	return func(r *options) {
-		r.validator = v
-	}
-}
-
-// WithNestedDir adds withNestedDir = true parameter. That means that we won't skip nested folders.
-func WithNestedDir() Opt {
-	return func(r *options) {
-		r.withNestedDir = true
-	}
-}
-
 // NewReader returns new Azure blob directory/file reader.
 // Must be called with WithDir(path string) or WithFile(path string) - mandatory.
 // Can be called with WithValidator(v validator) - optional.
@@ -153,6 +102,7 @@ func (r *Reader) streamDirectory(
 
 	pager := r.client.NewListBlobsFlatPager(r.containerName, &azblob.ListBlobsFlatOptions{
 		Prefix: &r.prefix,
+		Marker: &r.marker,
 	})
 
 	for pager.More() {
